requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain-google-genai
langchain-community
langchain-core
youtube-transcript-api
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic






